requirements.txt
feedparser==6.0.11
jinja2==3.1.4
python-dotenv==1.0.1
requests==2.32.3
